server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
apscheduler==3.8.1
schedule==1.1.0Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
requests==2.28.2
python-dotenv==1.0.0
APScheduler==3.10.1
schedule==1.1.0